server/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
google-genai==0.2.2
python-dotenv==1.0.1
pydantic==2.10.3
requests==2.32.3
